Adiciona metricas de count e histogram
