+++ src/Metrics.php
@@ -17,29 +17,6 @@ abstract class Metrics
         $this->prometheus = app('prometheus');
     }
 
-    public function count(int $count = 1)
-    {
-        try {
-            $counter = $this->prometheus->getCounter($this->metric);
-            $counter->incBy($count);
-        } catch (MetricNotFoundException $metricNotFoundException) {
-            $this->registerCounter();
-            $this->count($count);
-        }
-    }
-
-    public function observe()
-    {
-        try {
-            $this->prometheus
-                ->getHistogram($this->histogram)
-                ->observe($this->observe);
-        } catch (MetricNotFoundException $metricNotFoundException) {
-            $this->registerHistogram();
-            $this->observe();
-        }
-    }
-
     protected function registerCounter()
     {
         $this->prometheus->registerCounter($this->metric, $this->hint);
