--- src/PrometheusMetricsInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Arquivei\Prometheus;
-
-interface PrometheusMetricsInterface
-{
-    public function count(int $count = 1);
-    public function observe();
-}
